Allow the official RU cover page to be included.
The communication department wants to have a unified look. They
provide a unified cover page in Indesign. Add a small package
that prepends a pdf to the document. This method should address the
requirements of our communication departments.

The current drawback of this method is that the cover page is in
A4 on A4 stock. We also need a B5 version if the thesis will be
trimmed to B5 stock.